[test262] V8 Test262 importer/roller implementation
The script is supposed to be used both manually and by a bot in a recipe.

When used manually it will, by default, run 4 distinct phases to prepare the change (DEPS and other files), determine the impact of the change (newly failing tests), make further adjustments (update status file based on test results) and upload the change. The script expects that the devs run it in an environment where `gn args` was set for `out/Default` and a new branch was created.

When running in the context of a recipe, the script is invoked twice: once to prepare the change and another time to adjust the status file. Optionally, the recipe can trigger the compilation and tests to determine the impact of the change, between the 2 script invocation. The recipe also will take care of uploading the change.

Depends on changes in blinkpy https://crrev.com/c/4971212 .

Bug: v8:14002
Change-Id: I8c40ce591355529e4c0d0a4419660e30c9537fcd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4971838
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>
Cr-Commit-Position: refs/heads/main@{#90745}