[sandbox] Reference bytecode from InterpreterData via a trusted pointer
This is required for moving BytecodeArrays out of the sandbox. We may
also want to move InterpreterData out of the sandbox as it can often
occur in places where a BytecodeArray is normally expected. If we do
that, we can again reference BytecodeArrays from InterpreterData through
a regular tagged pointer.

Bug: chromium:1472252
Change-Id: I9fee01fa6273a9f185c0123f9fdc78876680c98d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5006377
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90752}