[maglev] Fix phi untagging skipping smi check
A CheckedSmiUntag can only be removed if there is no use of the phi
which requires the int32 to be within the smi value range.

For instance if we still have tagged uses downstream and they currently
assume UnsafeSmiTag, then we must stay in the smi range for correctness.

Bug: v8:7700, chromium:1499133
Change-Id: I340b48eb8d6e5eeb1cc6ca00fa5af0a4f2953513
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5004063
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90803}