[regalloc] Remove mid-tier register allocator
Background: As of earlier this year, the mid-tier regalloc was only
used as a fallback for very large Wasm functions, due to some
pathological compile times with the top-tier regalloc.
Since we have fixed the (known) issues in the top-tier register
allocator, the mid-tier regalloc is fully disabled via a Finch
experiment in Chrome M117 and M118 and by default in M119.
We can remove it now.

This CL
- removes the mid-tier regalloc implementation and build files,
- collapses the (now empty) class hierarchy of
`TopTierRegisterAllocationData` to just `RegisterAllocationData`,
- removes tests that required the mid-tier regalloc to reproduce.

Bug: v8:14082, chromium:1469533
Change-Id: I8dd8127eb7649dc96b8872f7cfd2ac244b8edac9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5009399
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90794}