[maglev] Add eager deopt registers to register snapshot
We proactively add the deopt registers to the register snapshot in
nodes that can eagerly deopt and do a deferred call.

Currently, this happens to these nodes:
- CheckedObjectToIndex
- CheckMapsWithMigration
- CheckValueEqualsString
- MaybeGrowAndEnsureWritableFastElements
- TryOnStackReplacement

In 4 of these nodes we were already adding the deopt registers
in an ad-hoc fashion. {CheckedObjectToIndex} currently does not
need, since it is *currently* guaranteed that the eager deopt
registers are subset of the live registers due to the lifetime
extension of deopt inputs. This could easily change in the future
though. This CL guarantees that we don't shoot ourselves in the foot.

Bug: chromium:1381335, chromium:1500857

Change-Id: I213f2f02861c5911bc0da3cfb1f5c8189928efbd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5013477
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90831}