[handles] Deprecate vector of handles in error messages
Some methods for constructing error messages take as parameters vectors
of handles to strings or objects. Such handles cannot be migrated to
direct handles, as the latter must be stack-allocated. This CL replaces
these vectors of handles with MemorySpan objects, as a first step in
the migration towards direct handles.

Bug: v8:13270
Change-Id: Ia730b3794fa8edf7b5ad07a747c3eb4690b42caf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5020707
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90894}