Add continuation preserved embedder data to Isolate
We want to move the continuation preserved embedder data from the
Context to the Isolate.
This CL is the first step on that path - adding a similar API surface
on the Isolate that could be called from the Blink side.

Future Blink-side CL would make use of this API instead of Context's.
Then a future v8 CL would remove the Context API.

Change-Id: Ia2cf99c42dffea0299bb7d95b4a6d371098ba58f
Bug: chromium:1351643
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5006444
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90996}