[sandbox] Split SFI::function_data into two fields
When the sandbox is enabled, we now have SFI::trusted_function_data and
SFI::function_data. Only one of the two can be "active" at a time. For
now, both fields are tagged pointers, but in a follow-up CL, the
trusted_function_data field will be converted into a trusted pointer
field so it can reference objects in trusted space.

Once we also move Code (and likely InterpreterData) objects into trusted
space, we could convert the trusted_function_data field into a "code"
field that contains a trusted pointer to the code (BytecodeArray, Code,
or InterpreterData) to use when executing the function.

Bug: chromium:1490564
Change-Id: I8e215c809a27604855bebb11f706e907950d6ccc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4922343
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90920}