Use checked truncation when computing branch offsets on arm64
This change hardens against certain bugs where we forget to check if
branch veneers are required on arm64. In these cases, the branch offset
will be too large, so we now perform a "checked" truncation that makes
sure that the input value fits into the output type.

Bug: chromium:1485829
Change-Id: Icb15c71345e24d57c7882d9de8044733391aedef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5029995
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90929}