[wasm] Mark fake Scripts as "shared cross origin"
Wasm modules don't have associated Script objects in the same way
as JS source code, and they reach the browser in different ways
(typically, via fetch() or embedded directly as array literals
etc., not via <script> tags).
For reasons of implementation uniformity, we still create fake Script
objects for them, so that e.g. the handling of exceptions can assume
that every frame that can throw has a Script.
Blink uses a particular attribute of Scripts, the IsSharedCrossOrigin
bit, to enforce the concept of "muted errors" [1], which prevents a
particular form of possible cross-origin information leakage by not
surfacing certain errors, in particular in the "onunhandledrejection"
callback, for non-shared scripts. Since the scenario this targets is
unrelated to Wasm, and since catching bugs in Wasm modules that are
instantiated and/or executed asynchronously in Promises is very useful,
this patch marks the fake Script objects we create for Wasm modules
as IsSharedCrossOrigin.

[1] https://html.spec.whatwg.org/#muted-errors

Change-Id: If35c1dec52fa02b6331cddad1dff7fa1336a94e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5013671
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90928}