[sandbox] Make InterpreterData a trusted object
In many places we use InterpreterData and BytecodeArray interchangeably
as InterpreterData is mostly just a wrapper around a BytecodeArray. For
moving BytecodeArrays out of the sandbox, we need to reference them via
trusted pointers from within the sandbox. Then, in order to keep union
fields that can reference either a BytecodeArray or an InterpreterData
object working, both of these objects must be referenced via a trusted
pointer and so need to be ExposedTrustedObjects living in trusted space.

Bug: chromium:1502461
Change-Id: I123915ddd1601f1ed0ee61bcdf195997694c643e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4976264
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90988}