[heap] Remove IsMainThread() uses in MainAllocator
Heap::IsMainThread() cannot be used on background threads because
it might race with V8::Locker there. It can be replaced with
MainAllocator::is_main_thread() in all but one use case.

In this case we want to use the the main thread scope even for
the GC main thread but MainAllocator::is_main_thread() would return
false in this case. Here we need to explicitly use
Heap::IsMainThread() after all. But during GC it is safe to use that
method from background threads because v8::Locker can't be used by the
embedder concurrently because the main thread is still occupied by the
GC.

Bug: chromium:1480975
Change-Id: I145b30360a8f4d5b05e55489e3bfb0b54de96583
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5033263
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90968}