[heap] Merge StartIncrementalMarkingIfLimitIsReached for all threads
Merge this method for main and background threads. The method now
gets the LocalHeap as an argument to decide whether it currently
runs on a main or background thread.

As soon as the hard limit is reached, incremental marking starts
right away on the main thread. However marking can not be started
on a background thread, so on such threads marking is lazily
started on the main thread both through a stack guard and posting a
task.

Note that this changes behavior in ConcurrentAllocator. So far this
Allocator always used the background thread version of this method.
And in that version marking was never started right away even on the
main thread.

This means that we now need to change order and invoke this method
before the allocation like we already do on the main thread in order
for black allocation to work. It also means that incremental marking
could start in situations where we didn't expect it so far (see changes
in tests for this).

Bug: chromium:1480975
Change-Id: I02460e07fc159a443e5e2ba00e48e988707cdcbb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5030192
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90994}