[sandbox] Introduce CodeWrapper type
Similar to the BytecodeWrapper type (introduced in crrev.com/c/4993365),
this is a HeapObject that contains a code pointer to the real Code
object. While the Code object will move out of the sandbox and into
trusted space, the wrapper object will remain inside the sandbox and can
therefore for example be used in situations where a reference to a Code
object needs to be stored in a collection of tagged values, such as a
FixedArray. There is a 1:1 relationship between Code and CodeWrapper.

Bug: chromium:1500887
Change-Id: I94d5a201f2ed02f91fc2d1adbd6554b5fa6f3cef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5017768
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91029}