[sandbox] Reference InterpreterData via SFI::trusted_function_data
InterpreterData is a trusted object and therefore should move into
trusted space. For that, it needs to be referenced from inside the
sandbox through a trusted pointer (so use a pointer table indirection).
This CL takes care of that for the reference from SharedFunctionInfo.

Bug: chromium:1502461
Change-Id: I6abd818ea040597667d487fc6501b7ef709dbbc0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5034002
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91034}