[wasm] Pass arguments directly to memory.init
Avoid setting up a stack buffer per C call as those stack buffers are
not shared and hence blow up the frame.

As this is the first C function with more than four parameters, we need
to implement logic to push remaining arguments to the stack on platforms
with less than six parameter registers.

R=thibaudm@chromium.org

Bug: v8:14357
Change-Id: Id315ce9b5c5127b4889523123a8a3e81c6086aa7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5040950
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91059}