[asm.js][turboshaft] eliminate unnecessary alignment masks
C++ code compiled to asm.js often contains memory accesses with
a computed address of the following pattern:
  addr = (C1 + x * C2 + y << C3) & -S
where S is the size of the access (often 4 bytes), and "& -S"
serves to align the value, but constants C1, C2, and (1 << C3)
are already multiples of S, so the alignment-masking is actually
a no-op.
This patch ports the required logic from TF to TS to allow the
compiler to recognize this fact and eliminate the unnecessary
masking.
This improves e.g. JetStream2's "float-mm" test by around 20%.

Bug: v8:14108
Change-Id: I4be6f0464b729c100928c3eb6858d9270e0d2f9d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5034123
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91041}