[release] Fix roll version retrieval in roll-merge-gerrit
Previously we tried to get the version from the currently rolled
commit from its tag, but that doesn't work. There's no method in
the Gerrit API to retrieve the tags on a commit. The commit/in
API includes all tags transitively a commit is included in. It'd
require additional calls for each such tag to verify which is
on the commit exactly.

Also gitiles doesn't work as the json api differs from the UI log
by just not including the tags. Local `git tag --points-at`
requires a local checkout and doesn't work with a remote repo.

The fix in this change is to use the version file content to
determine the version. The file is the source of truth anyways and
the tag usually just mimics the file anyways.

The CallGerritApi can unfortunately not be reused, as it only works
with json results, but the file API returns base64 encoded data.

No-Try: true
Bug: chromium:1445862
Change-Id: I31a6f74a3a214f767b03689fa2da5ebeae411df7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5048911
Reviewed-by: Liviu Rau <liviurau@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91081}