[wasm] Add frame state to calls to imports in wrapper inlining
An inlined js-to-wasm wrapper has to add a frame state to calls to
generate correct deopt information. For calls to imported functions,
this was not happening so far, which triggered a DCHECK. This CL adds a
frame state also to calls to imported functions.

Bug: chromium:1501914
Change-Id: I29f3da99cf87fbc012532c7cea6554a9dccb88fc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5050084
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91106}