[sandbox] Introduce IsolateForSandbox
This is a view of an Isolate that only exposed the sandbox-related
parts, in particular the various pointer tables stored on an Isolate.
Introducing this class achieves two things:

1. It allows accessing pointer table entries given either a Isolate* or
   a LocalIsolate*. Both of these are valid, but previously we didn't
   have proper APIs for using a LocalIsolate.

2. It makes it obvious when an isolate is only required for the sandbox.
   Previously, we sometimes names the parameter isolate_for_sandbox, but
   this was only done inconsistently.

As an added benefit, this CL also moves previously duplicated logic for
obtaining the appropriate external pointer table into the new
IsolateForSandbox class, thereby recuding code duplication.

With this change, we also forbid passing a nullptr Isolate into the
pointer accessor routines, which was previously allowed for code
pointers as these do not yet need an isolate. As such, we have to pass
an isolate into a number of additional function now. However, it is
expected that this will anyway be necessary in the future when the code
pointer table also becomes a per-Isolate table.

Bug: chromium:1503631
Change-Id: I0bdcacf5da6c5b13de828d72e9ec4bfb3313f9e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5043729
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91102}