[wasm][turboshaft] LoadOp: Add extra flag for trap_on_null
So far for protected instructions the tagged_base was used to decide
whether the load traps on null for wasm-gc instructions or traps on out
of bounds for linear wasm memory accesses.

The tagged_base is however also used to mark loads as "the offset has
to be reduced by one to remove the tag bit".
Lowering loads and stores in the LoadStoreSimplificationReducer removes
the tagged_base as it subtracts the tag bit already. It however still
needs to differentiate between the two different kinds of traps.

Bug: v8:12783
Change-Id: I4727d7bfaff5d7c6d041ff8cdeb4b91b52c7afbf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5047534
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91119}