[heap] Introduce GC flag for optimizing for battery
This CL introduces the new method Heap::ShouldOptimizeForBattery()
which can be used to check whether the GC should optimize for
battery now. Right now the only way this flag can be enabled is with
the new command line flag --optimize-gc-for-battery. In the future
this method will likely check a global battery mode state as well.

When this flag is enabled, concurrent marking and parallel scavenge
is restricted to a single thread. The idea here is that those
GC phases require per-object synchronization and are thus believed
to have higher overhead compared to phases that operate on a per-page
granularity. cppgc will be addressed in a follow-up CL.

Bug: v8:13842
Change-Id: Ifb1b044c9593f81054b23eea7fe3097760887deb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5046673
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91156}