[maglev] Fixed cached array length in forEach
The forEach inlining checks that the length of the array being iterated
doesn't decrease (so that the index starts going out of bounds), but
allows the length to increase (to avoid unnecessary deopts if the
callback appends to the array).

However, the array length is still allowed to _change_, which means that
the cached array length (for load elimination) must be a variable
(potentially a Phi) rather than keeping the original length.

Bug: chromium:1504982
Change-Id: Ib0f4349579c2a6848f7cff889bc8a52834225997
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5058590
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91158}