[sandbox] Align the trusted range at 4GB
We don't want trusted space to be allocated within the first 4GB of the
address space as this can be problematic both for security and
correctness, see the added comment. To achieve that, we now simply
require 4GB alignment of the trusted range. This may anyway later be
required if we want to also map the read-only space into the trusted
range so that the compressed references between RO objects are valid.

Bug: chromium:1504934, chromium:1473677
Change-Id: Ie671ae627bac20a9b93c632de19aaae55ca3a593
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5054646
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91161}