[heap] Switch to MainAllocator for shared space promotion
This CL switches the last use case of ConcurrentAllocator to
MainAllocator. The tricky thing about this particular use case is
that while the MainAllocator instance is used for GC, we are
actually not collecting this particular space in the case of
worker isolates and shared space.

This CL solves this by introducing a new method in_gc_for_space()
that only returns true if this LAB is used for GC and the space
is currently collected as well. Some use cases of in_gc() had to be
switched to in_gc_for_space() for correctness.

Alternatively one could argue that the allocator shouldn't even be
considered to be in a GC. However, we likely don't want shared GCs
to interrupt local GCs. So we would anyways need some kind of GC flag
for those allocators to prevent that from happening. There are
other smaller issues as well like that we don't have LocalHeaps for
GC threads that make this harder as well.

Bug: chromium:1480975
Change-Id: I7c707402e928e95cb092c4c183ee69cbae285f9d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5057489
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#91171}